Fix code in the documentation so that it compiles from the get go
Hello,
I am just starting a new project with Mapsui and copied the code above into a brand new empty project and it did not work. It could not find the Widgets even after adding a using Mapsui. The solution was to add the namespace in place as I did above.
I hope this helps and thank you for the fantastic project.
Jose